add nullish coalescing operator ?? renaming GasStation => GasSationList Co-authored-by: AlexF090 <AlexF090@users.noreply.github.com>
