+++ client/src/pages/LandingPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GasStation from '../components/GasStation.jsx';
+import GasStationList from '../components/GasStationList.jsx';
 import Header from '../components/Header.jsx';
 import stations from '../data/db';
 
@@ -7,7 +7,7 @@ function LandingPage({ title, toggleFavorite, favoriteIDs }) {
   return (
     <>
       <Header title={title} />
-      <GasStation stations={stations} toggleFavorite={toggleFavorite} favoriteIDs={favoriteIDs} />
+      <GasStationList stations={stations} toggleFavorite={toggleFavorite} favoriteIDs={favoriteIDs} />
     </>
   );
 }
